Change iter_parameters back and port unittests code to Python3

+++ python/paddle/fluid/tests/unittests/test_mean_iou.py
@@ -15,7 +15,7 @@
 from __future__ import division
 import unittest
 import numpy as np
-from op_test import OpTest
+from .op_test import OpTest
 
 
 def compute_mean_iou(predictions, labels, num_classes, in_wrongs, in_corrects,
@@ -80,7 +80,7 @@ class TestMeanIOUOp(OpTest):
             'InCorrects': in_corrects,
             'InMeanIou': in_mean_ious
         }
-        self.attrs = {'num_classes': long(self.num_classes)}
+        self.attrs = {'num_classes': int(self.num_classes)}
         mean_iou, out_wrong, out_correct = compute_mean_iou(
             predictions, labels, self.num_classes, in_wrongs, in_corrects,
             in_mean_ious)
